SCRUM-98 refactor: renaming PostCard to PublicPost
