+++ src/entities/users/ui/public/PublicPost/PublicPost.module.scss
